+++ src/components/Product/BestProducts.js
@@ -1,54 +1,8 @@
-import { useEffect, useState } from "react";
-import { getProductsList } from "../../services/panda-market-api";
+import useProductsFavorite from "../../hooks/useProductsFavorite";
 import ProductsList from "../../layout/ProductsList";
 
 const BestProducts = () => {
-  const [items, setItems] = useState([]);
-  const [loadingError, setLoadingError] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [pageSize, setPageSize] = useState(4);
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (window.matchMedia("(max-width: 767px)").matches) {
-        setPageSize(1);
-      } else if (window.matchMedia("(max-width: 1199px)").matches) {
-        setPageSize(2);
-      } else {
-        setPageSize(4);
-      }
-    };
-
-    handleResize();
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  const handleLoad = async () => {
-    const queryParams = {
-      pageSize,
-      orderBy: "favorite",
-    };
-    let result;
-    try {
-      setIsLoading(true);
-      setLoadingError(null);
-      result = await getProductsList(queryParams);
-    } catch (error) {
-      setLoadingError(error);
-    } finally {
-      setIsLoading(false);
-    }
-    const { list } = result;
-    setItems(list);
-  };
-
-  useEffect(() => {
-    handleLoad();
-  }, [pageSize]);
+  const { items, loadingError, isLoading } = useProductsFavorite();
 
   return (
     <>
